+++ assets/Assets.php
@@ -7,20 +7,36 @@
 
 namespace humhub\modules\popovervcard\assets;
 
+use humhub\modules\ui\view\components\View;
+use yii\helpers\Url;
 use yii\web\AssetBundle;
 
 class Assets extends AssetBundle
 {
 
     public $publishOptions = [
-        'forceCopy' => false
+        'forceCopy' => true
     ];
     public $sourcePath = '@popover-vcard/resources';
     public $css = [
-        'popover.css'
+        'humhub.vcard.popover.css'
     ];
     public $js = [
-        'popover.js'
+        'humhub.vcard.popover.js'
     ];
+
+    /**
+     * @param View $view
+     * @return void|AssetBundle
+     */
+    public static function register($view)
+    {
+        parent::register($view);
+
+        $view->registerJsConfig('vcard.popover', [
+            'delay' => 500,
+            'loadUrl' =>  Url::to(['/popover-vcard/index/load'])
+        ]);
+    }
 
 }